pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trest"
version = "1.4.0"
description = "Basic REST client with extended logging"
readme = "README.md"
authors = [{ name = "l0kifs", email = "l0kifs91@gmail.com" }]
license = { file = "LICENSE.md" }
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["rest", "client", "allure", "logging"]
dependencies = ["allure-pytest", "requests", "jto"]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = ["pip-tools", "pytest"]

[project.urls]
Homepage = "https://github.com/l0kifs/trest_client"
